Fix description metadata
There doesn't seem to be a "summary" option, but "description" and
"long_description" are different. Update the metadata to avoid warnings
about the unknown "summary" and so that when the release announce script
asks for the package description it gets the one line version instead of
the longer one.

Also, bump the build number as 1.4.13.1 has been released.

Change-Id: Ibe475d90c97f6e0c09be92c211924546d7d8cca1
Signed-off-by: Doug Hellmann <doug@doughellmann.com>